Finalized high level implementation for improved security on execution of console commands sent over the net. We now know where the command came from before executing (i.e. local or remote) and also the invocation method used (e.g. remote client console input, DED file etc). However, due to the fact that the remote invocation method MUST be sent over the net - any truly dangerous ccmds should be marked as CMDF_CLIENT. In order to implement this a new PKT_COMMAND2 type has been added which obsoletes the old type. Due to these measures any commands sent using the old PKT_COMMAND type will NEVER be executed.

+++ doomsday/engine/portable/include/con_decl.h
@@ -18,7 +18,7 @@
  *
  * You should have received a copy of the GNU General Public License
  * along with this program; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin St, Fifth Floor, 
+ * Foundation, Inc., 51 Franklin St, Fifth Floor,
  * Boston, MA  02110-1301  USA
  */
 
@@ -41,7 +41,7 @@
     { ccmd_t _c = { name, CCmd##fn, flags }; Con_AddCommand(&_c); }
 
 // A handy helper for declaring console commands.
-#define D_CMD(x) int CCmd##x(int src, int argc, char **argv)
+#define D_CMD(x) int CCmd##x(byte src, int argc, char **argv)
 
 /*
  * Macros for creating new console variables.
